{"schema":"decodex/commit/1","summary":"Split Decodex app panel layout support","authority":"XY-1115"}
