+++ apps/decodex-app/Sources/DecodexApp/AccountPanelPrivacy.swift
@@ -1,0 +1,4 @@
+enum AccountPrivacy {
+	static let hiddenValue = "hidden"
+	static let visibleValue = "visible"
+}
