+++ apps/decodex-app/Sources/DecodexApp/AccountPanelView.swift
@@ -523,126 +523,3 @@ struct AccountPanelView: View {
 		loginWindowState.isPresented = true
 	}
 }
-
-enum AccountPanelLayout {
-	static let accountListScrollSpace = "account-list-scroll"
-	static let screenVerticalMargin: CGFloat = 44
-	static let panelVerticalPadding: CGFloat = 18
-	static let sectionSpacing: CGFloat = 6
-	static let headerHeight: CGFloat = 28
-	static let accountSummaryHeight: CGFloat = 31
-	static let telemetryHorizontalPadding: CGFloat = 7
-	static let telemetryTopPadding: CGFloat = 7
-	static let telemetryBottomPadding: CGFloat = 2
-	static let telemetryVerticalPadding: CGFloat = telemetryTopPadding + telemetryBottomPadding
-	static let telemetryRowSpacing: CGFloat = 5
-	static let telemetryProfileHeight: CGFloat = 50
-	static let telemetryPoolHeight: CGFloat = 16
-	static let telemetryPoolMeasuredHeight: CGFloat = 29
-	static let noticeHeight: CGFloat = 44
-	static let minimumScrollableListHeight: CGFloat = 312
-
-	static func activeScreenVisibleHeight() -> CGFloat {
-		let mouseLocation = NSEvent.mouseLocation
-		let screen = NSScreen.screens.first { screen in
-			screen.frame.contains(mouseLocation)
-		} ?? NSScreen.main
-
-		return screen?.visibleFrame.height ?? 760
-	}
-}
-
-private struct AccountScrollOffsetPreferenceKey: PreferenceKey {
-	static let defaultValue: CGFloat = 0
-
-	static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-		value = nextValue()
-	}
-}
-
-private struct AccountListScrollIndicatorView: View {
-	let contentHeight: CGFloat
-	let viewportHeight: CGFloat
-	let scrollOffset: CGFloat
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		if contentHeight > viewportHeight + 1 {
-			ZStack(alignment: .top) {
-				Capsule(style: .continuous)
-					.fill(PanelPalette.secondaryText(colorScheme).opacity(0.12))
-					.frame(width: 3, height: viewportHeight)
-
-				Capsule(style: .continuous)
-					.fill(PanelPalette.secondaryText(colorScheme).opacity(colorScheme == .dark ? 0.42 : 0.34))
-					.frame(width: 3.5, height: thumbHeight)
-					.offset(y: thumbOffset)
-			}
-			.frame(width: 8, height: viewportHeight)
-			.allowsHitTesting(false)
-		}
-	}
-
-	private var thumbHeight: CGFloat {
-		max(30, viewportHeight * min(1, viewportHeight / max(contentHeight, 1)))
-	}
-
-	private var thumbOffset: CGFloat {
-		let maxScrollOffset = max(1, contentHeight - viewportHeight)
-		let maxThumbOffset = max(0, viewportHeight - thumbHeight)
-		let progress = min(1, max(0, scrollOffset / maxScrollOffset))
-
-		return maxThumbOffset * progress
-	}
-}
-
-private enum AccountPrivacy {
-	static let hiddenValue = "hidden"
-	static let visibleValue = "visible"
-}
-
-private func formatPanelTimestamp(_ value: String?) -> String? {
-	guard let value = panelTrimmed(value) else {
-		return nil
-	}
-
-	let date = parsePanelTimestamp(value)
-	guard let date else {
-		return value
-	}
-
-	let formatter = DateFormatter()
-	formatter.locale = Locale(identifier: "en_US_POSIX")
-	let calendar = Calendar(identifier: .gregorian)
-	formatter.dateFormat = calendar.component(.year, from: date) == calendar.component(.year, from: Date())
-		? "MMM d HH:mm"
-		: "MMM d yyyy HH:mm"
-	return formatter.string(from: date)
-}
-
-private func parsePanelTimestamp(_ value: String) -> Date? {
-	let fractionalFormatter = ISO8601DateFormatter()
-	fractionalFormatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-	if let date = fractionalFormatter.date(from: value) {
-		return date
-	}
-
-	return ISO8601DateFormatter().date(from: value)
-}
-
-private func compactLanePath(_ value: String?) -> String? {
-	guard var text = panelTrimmed(value), text.isEmpty == false else {
-		return nil
-	}
-
-	if let home = ProcessInfo.processInfo.environment["HOME"], home.isEmpty == false {
-		text = text.replacingOccurrences(of: home, with: "~")
-	}
-	if text.count <= 42 {
-		return text
-	}
-
-	let prefix = text.prefix(18)
-	let suffix = text.suffix(20)
-	return "\(prefix)...\(suffix)"
-}
